game/requirements.txt
wit==5.1.0
clarifai==2.4.2
Pillow==5.3.0
